# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Transfer Hub 360",
        "description": "Transfer Hub 360 API Documentation and Testing URL's",
        "contact": {
            "name": "Rana Raheel Tariq",
            "email": "ranaraheel@alfanovatech.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/companies/drivers?page={number}": {
            "get": {
                "tags": [
                    "CompanyDriver"
                ],
                "summary": "Get list of Drivers",
                "description": "Returns list of Drivers",
                "operationId": "getDriverList",
                "parameters": [
                    {
                        "name": "number",
                        "in": "path",
                        "description": "Page Number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": null
                                        },
                                        "data": {
                                            "type": "string",
                                            "example": "array of Drivers list"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/companies/drivers/create": {
            "post": {
                "tags": [
                    "CompanyDriver"
                ],
                "summary": "Create New Driver",
                "description": "Returns Driver data",
                "operationId": "storeDriver",
                "requestBody": {
                    "description": "I just fill Required Fields",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "identify_number",
                                    "full_name",
                                    "contact_number",
                                    "email",
                                    "password",
                                    "address",
                                    "gender",
                                    "status"
                                ],
                                "properties": {
                                    "identify_number": {
                                        "type": "string",
                                        "format": "identify_number",
                                        "example": "3663558DDSF"
                                    },
                                    "full_name": {
                                        "type": "string",
                                        "format": "full_name",
                                        "example": "Ahmet Ali"
                                    },
                                    "contact_number": {
                                        "type": "number",
                                        "format": "contact_number",
                                        "example": "00905340344609"
                                    },
                                    "email": {
                                        "type": "email",
                                        "format": "email",
                                        "example": "driver@yopmail.com"
                                    },
                                    "password": {
                                        "type": "password",
                                        "format": "password",
                                        "example": "dfdf323"
                                    },
                                    "thumbnail": {
                                        "type": "file",
                                        "format": "thumbnail",
                                        "example": ""
                                    },
                                    "address": {
                                        "type": "string",
                                        "format": "address",
                                        "example": "Zafar Mah. Cahmuriyat Cadde. Bachelivler"
                                    },
                                    "gender": {
                                        "type": "string",
                                        "format": "gender",
                                        "example": "Male"
                                    },
                                    "status": {
                                        "type": "in:Active,Deactive",
                                        "format": "status",
                                        "example": "Active"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Driver Successfully Created"
                                        },
                                        "data": {
                                            "type": "string",
                                            "example": "array of Driver Data"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "203": {
                        "description": "Validation Error response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validation error Message"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Content",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "!Something went wrong please try again later."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/companies/drivers/detail/{id}": {
            "get": {
                "tags": [
                    "CompanyDriver"
                ],
                "summary": "Get of Driver By Id",
                "description": "Get Driver Data by Driver Id",
                "operationId": "getdriverById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Driver Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": null
                                        },
                                        "data": {
                                            "type": "string",
                                            "example": "array of Driver"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Page Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Page Not Found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}